# adapters/handlers/availability.handler.go
package handlers

import (
	"github.com/labstack/echo/v4"
)

// @Summary Create availability for a diagnostic centre
// @Description Create a new availability slot for the diagnostic centre
// @Tags Availability
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param Authorization header string true "Bearer token"
// @Param RequestBody body domain.CreateAvailabilityDTO true "Availability information"
// @Success 201 {object} domain.AvailabilitySlot
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/availability [post]
func (handler *HTTPHandler) CreateAvailability(context echo.Context) error {
	return handler.service.CreateAvailability(context)
}

// GetAvailability retrieves availability slots for a diagnostic centre
// @Summary Get availability for a diagnostic centre
// @Description Get availability slots for the diagnostic centre, optionally filtered by day of week
// @Tags Availability
// @Accept json
// @Produce json
// @Param diagnostic_centre_id path string true "Diagnostic Centre ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param day_of_week query string false "Day of week" Enums(monday,tuesday,wednesday,thursday,friday,saturday,sunday)
// @Success 200 {array} domain.AvailabilitySlot "SUCCESS_RESPONSE"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/availability/{diagnostic_centre_id} [get]
func (handler *HTTPHandler) GetAvailability(context echo.Context) error {
	return handler.service.GetAvailability(context)
}

// @Summary Update availability for a diagnostic centre
// @Description Update an existing availability slot for the diagnostic centre
// @Tags Availability
// @Accept json
// @Produce json
// @Param diagnostic_centre_id path string true "Diagnostic Centre ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param day_of_week path string true "Day of week (monday, tuesday, etc.)"
// @Param RequestBody body domain.UpdateAvailabilityDTO true "Updated availability information"
// @Success 200 {object} domain.AvailabilitySlot "SUCCESS_RESPONSE"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/availability/{diagnostic_centre_id}/{day_of_week} [put]
func (h *HTTPHandler) UpdateAvailability(c echo.Context) error {
	return h.service.UpdateAvailability(c)
}

// @Summary Bulk update availability for a diagnostic centre
// @Description Update multiple availability slots for the diagnostic centre
// @Tags Availability
// @Accept json
// @Produce json
// @Param diagnostic_centre_id path string true "Diagnostic Centre ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param RequestBody body domain.UpdateManyAvailabilityDTO true "Updated availability information"
// @Success 200 {array} domain.AvailabilitySlot "SUCCESS_RESPONSE"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/availability/{diagnostic_centre_id} [put]
func (handler *HTTPHandler) UpdateManyAvailability(context echo.Context) error {
	return handler.service.UpdateManyAvailability(context)
}

// @Summary Delete availability for a diagnostic centre
// @Description Delete an availability slot for the diagnostic centre
// @Tags Availability
// @Accept json
// @Produce json
// @Param diagnostic_centre_id path string true "Diagnostic Centre ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param day_of_week path string true "Day of week (monday, tuesday, etc.)"
// @Success 204 "No Content"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/availability/{diagnostic_centre_id}/{day_of_week} [delete]
func (h *HTTPHandler) DeleteAvailability(c echo.Context) error {
	return nil
}