# swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Diagnoxix API",
        "title": "Diagnoxix",
        "contact": {},
        "version": "1.0"
    },
    "host": "diagnoxix.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v1/availability": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new availability slot for the diagnostic centre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Create availability for a diagnostic centre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Availability information",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateAvailabilityDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.AvailabilitySlot"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/availability/{diagnostic_centre_id}": {
            "get": {
                "description": "Get availability slots for the diagnostic centre, optionally filtered by day of week",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Get availability for a diagnostic centre",
                "parameters": [
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Diagnostic Centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "monday",
                            "tuesday",
                            "wednesday",
                            "thursday",
                            "friday",
                            "saturday",
                            "sunday"
                        ],
                        "type": "string",
                        "description": "Day of week",
                        "name": "day_of_week",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.AvailabilitySlot"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "put": {
                "description": "Update multiple availability slots for the diagnostic centre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Bulk update availability for a diagnostic centre",
                "parameters": [
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Diagnostic Centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated availability information",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateManyAvailabilityDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.AvailabilitySlot"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/availability/{diagnostic_centre_id}/{day_of_week}": {
            "put": {
                "description": "Update an existing availability slot for the diagnostic centre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Update availability for a diagnostic centre",
                "parameters": [
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Diagnostic Centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Day of week (monday, tuesday, etc.)",
                        "name": "day_of_week",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated availability information",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateAvailabilityDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "$ref": "#/definitions/domain.AvailabilitySlot"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an availability slot for the diagnostic centre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Availability"
                ],
                "summary": "Delete availability for a diagnostic centre",
                "parameters": [
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Diagnostic Centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Day of week (monday, tuesday, etc.)",
                        "name": "day_of_week",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.AvailabilitySlot": {
            "type": "object",
            "properties": {
                "break_time": {
                    "description": "minutes",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "string"
                },
                "diagnostic_centre_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_appointments": {
                    "type": "integer"
                },
                "slot_duration": {
                    "description": "minutes",
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.CreateAvailabilityDTO": {
            "type": "object",
            "required": [
                "diagnostic_centre_id",
                "slots"
            ],
            "properties": {
                "diagnostic_centre_id": {
                    "type": "string"
                },
                "slots": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.Slots"
                    }
                }
            }
        },
...
        "domain.Slots": {
            "type": "object",
            "required": [
                "break_time",
                "day_of_week",
                "end_time",
                "max_appointments",
                "slot_duration",
                "start_time"
            ],
            "properties": {
                "break_time": {
                    "description": "minutes",
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                },
                "day_of_week": {
                    "type": "string",
                    "enum": [
                        "monday",
                        "tuesday",
                        "wednesday",
                        "thursday",
                        "friday",
                        "saturday",
                        "sunday"
                    ],
                    "example": "monday"
                },
                "end_time": {
                    "type": "string",
                    "example": "2025-08-27T17:00:00.000Z"
                },
                "max_appointments": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 20
                },
                "slot_duration": {
                    "description": "minutes",
                    "type": "integer",
                    "minimum": 1,
                    "example": 30
                },
                "start_time": {
                    "type": "string",
                    "example": "2025-08-27T09:00:00.000Z"
                }
            }
        },
...
        "domain.UpdateAvailabilityDTO": {
            "type": "object",
            "properties": {
                "break_time": {
                    "description": "minutes",
                    "type": "integer",
                    "minimum": 0
                },
                "end_time": {
                    "type": "string"
                },
                "max_appointments": {
                    "type": "integer",
                    "minimum": 0
                },
                "slot_duration": {
                    "description": "minutes",
                    "type": "integer",
                    "minimum": 1
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateManyAvailabilityDTO": {
            "type": "object",
            "required": [
                "slots"
            ],
            "properties": {
                "slots": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/domain.UpdateManyAvailabilitySlot"
                    }
                }
            }
        },
        "domain.UpdateManyAvailabilitySlot": {
            "type": "object",
            "required": [
                "day_of_week",
                "diagnostic_centre_id"
            ],
            "properties": {
                "break_time": {
                    "description": "minutes",
                    "type": "integer",
                    "minimum": 0
                },
                "day_of_week": {
                    "type": "string"
                },
                "diagnostic_centre_id": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "max_appointments": {
                    "type": "integer",
                    "minimum": 1
                },
                "slot_duration": {
                    "description": "minutes",
                    "type": "integer",
                    "minimum": 1
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "handlers.BAD_REQUEST": {
            "description": "BAD_REQUEST response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "BAD_REQUEST"
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.INTERNAL_SERVER_ERROR": {
            "description": "INTERNAL_SERVER_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "INTERNAL_SERVER_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Internal Server Error"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.NOT_FOUND_ERROR": {
            "description": "NOT_FOUND_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "NOT_FOUND_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Data not found"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.UNAUTHORIZED_ERROR": {
            "description": "UNAUTHORIZED_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "UNAUTHORIZED_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Unauthorized to perform this operation"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    }
}